# frontend/shared/openapi.json
{
  "components": {
    "schemas": {
...
      "BranchPatch": {
        "additionalProperties": false,
        "properties": {
          "is_archived": {
            "title": "Is Archived",
            "type": "boolean"
          }
        },
        "required": [
          "is_archived"
        ],
        "title": "BranchPatch",
        "type": "object"
      },
...
      "ChatDeleteBody": {
        "additionalProperties": false,
        "properties": {
          "confirm_irreversible": {
            "title": "Confirm Irreversible",
            "type": "boolean"
          },
          "request_id": {
            "format": "uuid",
            "title": "Request Id",
            "type": "string"
          }
        },
        "required": [
          "confirm_irreversible",
          "request_id"
        ],
        "title": "ChatDeleteBody",
        "type": "object"
      },
      "ChatListPage": {
        "additionalProperties": false,
        "properties": {
          "has_more": {
            "default": false,
            "title": "Has More",
            "type": "boolean"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/ChatSummary"
            },
            "title": "Items",
            "type": "array"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "required": [
          "items"
        ],
        "title": "ChatListPage",
        "type": "object"
      },
...
      "ChatPatch": {
        "additionalProperties": false,
        "properties": {
          "title": {
            "maxLength": 100,
            "minLength": 1,
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "title"
        ],
        "title": "ChatPatch",
        "type": "object"
      },
      "ChatSummary": {
        "additionalProperties": false,
        "properties": {
          "active_branch_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Active Branch Id"
          },
          "active_branch_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Active Branch Name"
          },
          "branch_count": {
            "minimum": 0.0,
            "title": "Branch Count",
            "type": "integer"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "last_message_preview": {
            "default": "",
            "title": "Last Message Preview",
            "type": "string"
          },
          "message_count": {
            "minimum": 0.0,
            "title": "Message Count",
            "type": "integer"
          },
          "storage_version": {
            "const": 2,
            "default": 2,
            "title": "Storage Version",
            "type": "integer"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "title",
          "active_branch_id",
          "branch_count",
          "message_count",
          "created_at",
          "updated_at"
        ],
        "title": "ChatSummary",
        "type": "object"
      },
      "ConversationTreeView": {
        "additionalProperties": false,
        "properties": {
          "branches": {
            "items": {
              "$ref": "#/components/schemas/BranchSummary"
            },
            "title": "Branches",
            "type": "array"
          },
          "chat": {
            "$ref": "#/components/schemas/ChatSummary"
          }
        },
        "required": [
          "chat",
          "branches"
        ],
        "title": "ConversationTreeView",
        "type": "object"
      },
...
      "ForkReason": {
        "enum": [
          "root",
          "rewind_continue",
          "edit_resend",
          "regenerate",
          "concurrent_send"
        ],
        "title": "ForkReason",
        "type": "string"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "MessageFeedbackBody": {
        "additionalProperties": false,
        "properties": {
          "branch_id": {
            "format": "uuid",
            "title": "Branch Id",
            "type": "string"
          },
          "rating": {
            "$ref": "#/components/schemas/MessageFeedbackRating"
          }
        },
        "required": [
          "branch_id",
          "rating"
        ],
        "title": "MessageFeedbackBody",
        "type": "object"
      },
      "MessageFeedbackRating": {
        "enum": [
          "like",
          "dislike"
        ],
        "title": "MessageFeedbackRating",
        "type": "string"
      },
      "MessageFeedbackView": {
        "additionalProperties": false,
        "properties": {
          "message_id": {
            "format": "uuid",
            "title": "Message Id",
            "type": "string"
          },
          "rating": {
            "$ref": "#/components/schemas/MessageFeedbackRating"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "message_id",
          "rating",
          "updated_at"
        ],
        "title": "MessageFeedbackView",
        "type": "object"
      },
      "MessagePathPage": {
        "additionalProperties": false,
        "properties": {
          "branch_id": {
            "format": "uuid",
            "title": "Branch Id",
            "type": "string"
          },
          "chat_id": {
            "title": "Chat Id",
            "type": "integer"
          },
          "has_more": {
            "default": false,
            "title": "Has More",
            "type": "boolean"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/ChatMessageView"
            },
            "title": "Items",
            "type": "array"
          },
          "next_before": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Before"
          }
        },
        "required": [
          "chat_id",
          "branch_id",
          "items"
        ],
        "title": "MessagePathPage",
        "type": "object"
      },
...
      "TurnAction": {
        "enum": [
          "append",
          "rewind_continue",
          "edit_resend"
        ],
        "title": "TurnAction",
        "type": "string"
      },
      "TurnCommand": {
        "additionalProperties": false,
        "description": "创建一轮对话的幂等命令。",
        "properties": {
          "action": {
            "$ref": "#/components/schemas/TurnAction",
            "default": "append"
          },
          "branch_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Branch Id"
          },
          "client_request_id": {
            "format": "uuid",
            "title": "Client Request Id",
            "type": "string"
          },
          "content": {
            "default": "",
            "maxLength": 20000,
            "title": "Content",
            "type": "string"
          },
          "derived_from_message_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Derived From Message Id"
          },
          "parent_message_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent Message Id"
          }
        },
        "required": [
          "client_request_id"
        ],
        "title": "TurnCommand",
        "type": "object"
      },
      "TurnCreationResult": {
        "additionalProperties": false,
        "properties": {
          "assistant_message_id": {
            "format": "uuid",
            "title": "Assistant Message Id",
            "type": "string"
          },
          "branch_created": {
            "title": "Branch Created",
            "type": "boolean"
          },
          "branch_id": {
            "format": "uuid",
            "title": "Branch Id",
            "type": "string"
          },
          "chat_id": {
            "title": "Chat Id",
            "type": "integer"
          },
          "fork_reason": {
            "$ref": "#/components/schemas/ForkReason"
          },
          "generation_id": {
            "format": "uuid",
            "title": "Generation Id",
            "type": "string"
          },
          "idempotent_replay": {
            "default": false,
            "title": "Idempotent Replay",
            "type": "boolean"
          },
          "user_message_id": {
            "format": "uuid",
            "title": "User Message Id",
            "type": "string"
          }
        },
        "required": [
          "chat_id",
          "branch_id",
          "user_message_id",
          "assistant_message_id",
          "generation_id",
          "branch_created",
          "fork_reason"
        ],
        "title": "TurnCreationResult",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "基于 LLM 的智能对话匹配系统原型",
    "title": "智能生育匹配系统 - 对话智能体",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/chats": {
      "get": {
        "operationId": "list_chats_v2_api_chats_get",
        "parameters": [
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "minimum": 1,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatListPage"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Chats V2",
        "tags": [
          "chats-v2"
        ]
      }
    },
    "/api/chats/turns": {
      "post": {
        "operationId": "create_new_chat_turn_api_chats_turns_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TurnCommand"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TurnCreationResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create New Chat Turn",
        "tags": [
          "chats-v2"
        ]
      }
    },
    "/api/chats/{chat_id}": {
      "delete": {
        "operationId": "delete_chat_v2_api_chats__chat_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "chat_id",
            "required": true,
            "schema": {
              "title": "Chat Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatDeleteBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Delete Chat V2",
        "tags": [
          "chats-v2"
        ]
      },
      "get": {
        "operationId": "get_chat_tree_api_chats__chat_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "chat_id",
            "required": true,
            "schema": {
              "title": "Chat Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConversationTreeView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Chat Tree",
        "tags": [
          "chats-v2"
        ]
      },
      "patch": {
        "operationId": "patch_chat_api_chats__chat_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "chat_id",
            "required": true,
            "schema": {
              "title": "Chat Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatPatch"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Patch Chat",
        "tags": [
          "chats-v2"
        ]
      }
    },
    "/api/chats/{chat_id}/branches/{branch_id}": {
      "patch": {
        "operationId": "patch_branch_api_chats__chat_id__branches__branch_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "chat_id",
            "required": true,
            "schema": {
              "title": "Chat Id",
              "type": "integer"
            }
          },
          {
            "in": "path",
            "name": "branch_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Branch Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BranchPatch"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Patch Branch",
        "tags": [
          "chats-v2"
        ]
      }
    },
    "/api/chats/{chat_id}/branches/{branch_id}/messages": {
      "get": {
        "operationId": "get_branch_messages_api_chats__chat_id__branches__branch_id__messages_get",
        "parameters": [
          {
            "in": "path",
            "name": "chat_id",
            "required": true,
            "schema": {
              "title": "Chat Id",
              "type": "integer"
            }
          },
          {
            "in": "path",
            "name": "branch_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Branch Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "before",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Before"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "minimum": 1,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessagePathPage"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Branch Messages",
        "tags": [
          "chats-v2"
        ]
      }
    },
    "/api/chats/{chat_id}/turns": {
      "post": {
        "operationId": "create_chat_turn_api_chats__chat_id__turns_post",
        "parameters": [
          {
            "in": "path",
            "name": "chat_id",
            "required": true,
            "schema": {
              "title": "Chat Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TurnCommand"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TurnCreationResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Chat Turn",
        "tags": [
          "chats-v2"
        ]
      }
    },
...
    "/api/messages/{message_id}/feedback": {
      "delete": {
        "operationId": "delete_message_feedback_api_messages__message_id__feedback_delete",
        "parameters": [
          {
            "in": "path",
            "name": "message_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Message Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Delete Message Feedback",
        "tags": [
          "chats-v2"
        ]
      },
      "put": {
        "operationId": "set_message_feedback_api_messages__message_id__feedback_put",
        "parameters": [
          {
            "in": "path",
            "name": "message_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Message Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MessageFeedbackBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageFeedbackView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Set Message Feedback",
        "tags": [
          "chats-v2"
        ]
      }
    },
    "/api/messages/{message_id}/match-results": {
      "get": {
        "operationId": "get_message_match_results_api_messages__message_id__match_results_get",
        "parameters": [
          {
            "in": "path",
            "name": "message_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Message Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "minimum": 1,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Message Match Results",
        "tags": [
          "chats-v2"
        ]
      }
    },
...
  }
}